Move `HTTPStatus` into its own file.
